rename package libnemo-extension1a -> libnemo-extension1
libnemo-extension1a was a left over of a workaround introduced by the Debian
packaging for Nautilus. This workaround is completely useless for packaging
of nemo and actually is harmful since it makes the packaging less standard
and rise several of warning/error from ther QA checks. Kept
libnemo-extension1a as transitional package.