Fix custom style classes for nemo-cell-renderer-disk.c:
On most themes, the way we had it worked just fine, however
on Adwaita, or, it seems, any theme using a compiled .gresource
file, we would get a 100% cpu draw on Nemo.

Applying the custom styles from our own .gresource file and loading
it once only on startup solves this issue.