Properly render the thumbnail when viewing a folder in icon mode and the file icons are smaller than the view-specified size.
To reproduce -

Enter /usr/share/icons/Mint-X/apps/16

You get lots of log spam, and the emblem for sym-linked files does
not scale with the icon size, or doesn't show at all, in the case
of a 16px icon, it is trying to place 24px emblem on top of it,
which doesn't work.  It works improperly for larger icon sizes -
the emblem renders, but is so large as to cover the file icon almost
completely.

This patch checks the *actual* size of the pixmap before trying to
apply the emblems, and adjusts the desired size accordingly.  As a
result, the emblems apply and are scaled appropriately to the icon
size.