When using panes, don't reset the selection each time the inactive pane is made active.
For example, you select some files in the left pane, then switch to the
right pane, do some navigation, then click back in the original (now
inactive) pane.

With the current behavior, the original selection is cleared as the pane
is made active again.  This patch prevents this, using the first click back
into the inactive pane to make the pane active *only*, not affecting the
existing selection.