Rename function `nemo_window_pane_close_slot'
The header for nemo-window-pane.h exposes the functions
`nemo_window_pane_close_slot' as well as `nemo_window_pane_slot_close'.
This naming is just unfortunate. To try and clarify the situation a bit,
rename `nemo_window_pane_close_slot' to
`nemo_window_pane_remove_slot_unsafe' as the latter really just removes
the slot from the GtkNotebook. The `*_unsafe' portion of the name refers
to the fact that it is not safe to call the function with the active
slot of a pane as this might crash Nemo.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>