Don't write to the legacy bookmarks file
The legacy bookmarks file should only be used when loading bookmarks and
no file is found in the default location. By letting
`nemo_bookmark_list_get_file()` fall back to the legacy file if the
default one isn't found, we would previously write to the legacy file
when bookmarks are changed instead of creating a new one in the default
location.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>